pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "system-operations-cli"
version = "0.1.0"
description = "A comprehensive Python-based CLI for managing distributed systems, deployments, monitoring, and infrastructure operations."
readme = "README.md"
license = "MIT"
authors = [
    { name = "jsamuelsen11", email = "jsamuelsen11@gmail.com" },
]
maintainers = [
    { name = "jsamuelsen11", email = "jsamuelsen11@gmail.com" },
]
requires-python = ">=3.14.3"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Systems Administration",
    "Topic :: System :: Monitoring",
    "Topic :: Utilities",
]
keywords = [
    "cli",
    "devops",
    "deployment",
    "monitoring",
    "kubernetes",
    "infrastructure",
    "distributed-systems",
    "automation",
]
dependencies = [
    "click>=8.3.1",
    "typer>=0.24.1",
    "rich>=14.3.3",
    "inquirerpy>=0.3.4",
    "sh>=2.2.2",
    "plumbum>=1.10.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "structlog>=25.5.0",
    "cryptography>=46.0.5",
    "jsonschema>=4.26.0",
    "psutil>=7.2.2",
    "pluggy>=1.6.0",
    "tenacity>=9.1.4",
    "textual>=8.0.0",
]

[project.optional-dependencies]
cli = [
    "cloup>=3.0.8",
    "cement>=3.0.14",
    "blessed>=1.30.0",
    "prompt-toolkit>=3.0.52",
]
kubernetes = [
    "kubernetes>=35.0.0",
    "kr8s>=0.20.15",
    "ruamel-yaml>=0.19.1",
]
monitoring = [
    "prometheus-client>=0.24.1",
    "grafana-api>=1.0.3",
    "elasticsearch>=9.3.0",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
]
vault = [
    "hvac>=2.4.0",
    "boto3>=1.42.54",
    "azure-keyvault-secrets>=4.10.0",
    "azure-identity>=1.25.2",
    "google-cloud-secret-manager>=2.26.0",
    "pyjwt>=2.11.0",
]
all = [
    "system-operations-cli[cli,kubernetes,monitoring,vault]",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.2",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "sphinx>=9.1.0",
    "sphinx-rtd-theme>=3.1.0",
    "types-pyyaml>=6.0.12.20250915",
    "testcontainers>=4.14.1",
    "respx>=0.22.0",
]

[project.urls]
Homepage = "https://github.com/Recipe-Web-App/system-control?tab=readme-ov-file#system-control-cli"
Documentation = "https://github.com/Recipe-Web-App/system-control/tree/main/docs"
Repository = "https://github.com/Recipe-Web-App/system-control"
"Bug Tracker" = "https://github.com/Recipe-Web-App/system-control/issues"
Changelog = "https://github.com/Recipe-Web-App/system-control/blob/main/CHANGELOG.md"

[project.scripts]
ops = "system_operations_manager.cli.main:cli"
test-cov = "scripts.test_scripts:coverage"
test-e2e = "scripts.test_scripts:e2e"
test-integration = "scripts.test_scripts:integration"
test-unit = "scripts.test_scripts:unit"

[project.entry-points."system_operations_manager.plugins"]
core = "system_operations_manager.plugins.core:CorePlugin"
kong = "system_operations_manager.plugins.kong:KongPlugin"
kubernetes = "system_operations_manager.plugins.kubernetes:KubernetesPlugin"

[tool.hatch.build.targets.wheel]
packages = ["src/system_operations_manager", "scripts"]

[tool.ruff]
line-length = 100
target-version = "py314"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # Pyflakes (logic/syntax bugs)
    "I",     # isort (import sorting)
    "UP",    # pyupgrade (modernizes syntax)
    "B",     # flake8-bugbear (common design pitfalls)
    "SIM",   # flake8-simplify (identifies complex logic)
    "PTH",   # flake8-use-pathlib (modern file paths)
    "RUF",   # Ruff-specific internal rules
    "ASYNC", # flake8-async (crucial for distributed apps)
    "T20",   # flake8-print (flags print statements)
]

ignore = [
    "E501",   # Ignore line-length (the formatter handles this automatically)
    "D100",   # Missing docstring in public module
    "D104",   # Missing docstring in public package
    "B008",   # Do not perform function calls in argument defaults (Too strict for Typer/Click)
    "RUF012", # Mutable class attributes should be annotated with `typing.ClassVar`
]

[tool.ruff.lint.per-file-ignores]
"src/cli/main.py" = ["T201"] # Main can use print statements
"tests/**/*" = ["S101", "D"] # Tests can use asserts and don't need docstrings

[tool.ruff.lint.isort]
known-first-party = ["system_operations_manager"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.14"
show_error_codes = true
pretty = true
strict = true
warn_return_any = true
warn_unused_ignores = true
no_implicit_optional = true
strict_optional = true
ignore_missing_imports = true
disallow_untyped_decorators = false

[[tool.mypy.overrides]]
module = [
    "sh.*",
    "plumbum.*",
    "inquirerpy.*",
    "kubernetes.*",
    "hvac.*",
    "prometheus_client.*",
    "pluggy.*",
    "testcontainers.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.integration.plugins.kong.conftest"
disallow_subclassing_any = false

[tool.pytest.ini_options]
minversion = "7.0"
addopts = ["-ra", "--strict-markers", "--strict-config"]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "slow: Slow running tests",
    "kubernetes: Tests requiring Kubernetes",
    "vault: Tests requiring Vault",
    "kong: Tests requiring Kong container",
    "kong_enterprise: Tests requiring Kong Enterprise",
    "requires_license: Tests requiring Kong Enterprise license for write operations",
    "requires_db: Tests requiring Kong database mode (not DB-less)",
    "requires_konnect: Tests requiring Konnect control plane connection",
]
filterwarnings = [
    # Suppress testcontainers internal deprecation (not in our code)
    "ignore:The @wait_container_is_ready decorator is deprecated:DeprecationWarning",
]

# Coverage configuration
[tool.coverage.run]
source = ["src/system_operations_manager"]
omit = ["*/tests/*", "*/test_*.py", "*/__pycache__/*", "*/venv/*", "*/.venv/*"]

[tool.coverage.report]
fail_under = 90
precision = 2
show_missing = true
skip_covered = true
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.bandit]
exclude_dirs = ["tests", "test_*"]
skips = ["B101", "B601"]
